update clean schedule to work across datasets
