+++ .env.sample
@@ -5,4 +5,4 @@ WORKSPACE_LOCATION=Path/to/workspace_files
 SHARED_LOCATION=Path/to/shared_files
 INPUT_LOCATION_EXT=Path/to/external/input_files
 INPUT_LOCATION_903=Path/to/903/input_files
-DATASET=Dataset code e.g. ssda903, cin
+ALLOWED_DATASETS=Dataset codes separated by comma (no spaces) e.g. ssda903,cin
